moved what was left in makemaps into webpage.py
